rebuild-system.sh: Be smarter about comparing symlinks and remote deploys
Instead of forwarding command-line arguments to nixos-rebuild, always
perform a fully local "nixos-rebuild ... build", and print the command
needed to deploy. The local build is needed in order for the ./result
symlink to be created.

Also we don't want to build with --target-host as that copies the closure
to the remote system without enabling it, as wasteful copy in case we
don't actually do the switch.

When run with no hostname, default to the local host.

Finally, keep separate symlinks per-host, so that the old/new comparison
will show relevant symlinks (instead of comparing apples to pears).